requirements.txt
gunicorn
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jws==0.1.3
lxml==4.5.0
MarkupSafe==1.1.1
networkx==2.4
nltk==3.5
numpy==1.18.2
oauth2client==4.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyPDF2==1.26.0
Pyrebase4==4.3.0
python-docx==0.8.10
python-jwt==2.0.1
regex==2020.4.4
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
soupsieve==2.0
stripe==2.45.0
tqdm==4.45.0
urllib3==1.25.9
validate-email==1.3
Werkzeug==1.0.1
